# DEV/DEV-03-openapi.yaml
openapi: 3.0.3
info:
  title: USTAPRO / SOZO Platform API
  version: 0.1.0
  description: |
    Контракт API платформы USTAPRO (Ташкент, ремонтно-сервисная платформа).
    Скелет для параллельной разработки фронтенда и бэкенда (монолит + PostgreSQL).

    Принципы:
    - Все изменения статуса заявки — только через POST /orders/{id}/transitions;
      недопустимый переход возвращает 409 с машиночитаемым кодом (PRD-05, раздел 1).
    - Суммы везде в тийинах (BIGINT, 1 сум = 100 тийин), поле `*_tiyin`.
    - Идемпотентность мобильных операций мастера — заголовок/поле `client_op_uuid`.
    - Аутентификация — JWT Bearer (access 15 мин / refresh 30 дней), вход по OTP-SMS.
    - Доменные события (order.*, billing.*, pricelist.release_activated) публикуются
      во внутреннюю шину; см. x-events в конце файла.
servers:
  - url: https://api.sozo.uz/v1
    description: Продакшн (домен sozo.uz — DEV-05 п.3; при смене бренда по итогам нейминга заменить единожды здесь)
  - url: https://staging-api.sozo.uz/v1
    description: Стейджинг
...
security:
  - bearerAuth: []
paths:
...
  /billing/invoices:
    get:
      tags: [billing]
      operationId: listInvoices
      summary: Счета-фактуры (для бухгалтера / руководителя организации)
      parameters:
        - name: organization_id
          in: query
          schema: { type: string, format: uuid }
        - name: status
          in: query
          schema:
            type: string
            enum: [draft, sent, paid, overdue]
      responses:
        '200':
          description: СФ
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Invoice'
...
  /billing/transactions:
    get:
      tags: [billing]
      operationId: listTransactions
      summary: Проводки (read-only; правки только сторно через админ-операции)
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PerPage'
        - name: order_id
          in: query
          schema: { type: string, format: uuid }
        - name: account_code
          in: query
          schema: { type: string }
        - name: date_from
          in: query
          schema: { type: string, format: date }
      responses:
        '200':
          description: Проводки двойной записи
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Transaction'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        Access-токен, полученный через /auth/verify (+ выбор контекста роли).
        В claims: user_id, role, organization_id?, location_id?, tenant_id.
  parameters:
...
    Page:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
    PerPage:
      name: per_page
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 25 }
...
  schemas:
...
    Invoice:
      type: object
      properties:
        id: { type: string, format: uuid }
        number: { type: string, example: 'INV-2026-000045' }
        organization_id: { type: string, format: uuid }
        kind:
          type: string
          enum: [subscription, overlimit, one_off]
          x-russian-label:
            subscription: 'Абонентка'
            overlimit: 'Сверхлимит + материалы'
            one_off: 'Разовая'
        period_start: { type: string, format: date }
        period_end: { type: string, format: date }
        total_tiyin: { type: integer, format: int64 }
        vat_tiyin:
          type: integer
          format: int64
          description: Выделенный НДС (0 для неплательщиков)
        status:
          type: string
          enum: [draft, sent, paid, overdue]
        edo_ref: { type: string, nullable: true }
        pdf_url: { type: string, format: uri, nullable: true }
        lines:
          type: array
          items:
            type: object
            properties:
              order_id: { type: string, format: uuid, nullable: true }
              description: { type: string }
              amount_tiyin: { type: integer, format: int64 }
              vat_tiyin: { type: integer, format: int64 }
    Transaction:
      type: object
      description: Проводка двойной записи (read-only через API)
      properties:
        id: { type: string, format: uuid }
        operation_id:
          type: string
          format: uuid
          description: Группа проводок одной операции (Σ дебетов = Σ кредитов)
        type: { type: string, example: 'subscription_writeoff' }
        debit_account: { type: string, example: 'subscription_advances' }
        credit_account: { type: string, example: 'revenue_works' }
        amount_tiyin: { type: integer, format: int64 }
        order_id: { type: string, format: uuid, nullable: true }
        invoice_id: { type: string, format: uuid, nullable: true }
        is_storno: { type: boolean }
        reversed_tx_id: { type: string, format: uuid, nullable: true }
        created_at: { type: string, format: date-time }
...